Remove 1 VA Domain
VA has stated: "The site does appear to be affiliated with VA, but it’s not.  An employee who works adjacent to the site administrator originally brought the site to our attention.  It took some time to establish reliable communication with the site administrator, but they replied in writing and stated the site is not owned or funded by VA." Removing since the domain is no longer owned by nor operated on behalf of VA.