Remove 2 HHS Domains
HHS has requested that we remove the domains “ahrq-meps.com” and “sops-database.org” as they are no longer owned or affiliated with AHRQ and may be removed from monitoring/tracking for the purposes of BOD 18-01. I verified that this domain is no longer active or owned by HHS via whois, dig, and nslookup.