Add TREAS domain
TREAS has requested that we add tftc-istehdaf.org so that it shows up in their Trustymail and HTTPS reports. I verified it is not currently being picked up in the sources gathered by double checking for its presence in their report. I also received confirmation that this domain is owned and operated by Treasury.